Move the Scope; some clean-up needed
