+++ mypy/semanal_pass3.py
@@ -18,7 +18,7 @@ from mypy.nodes import (
     Node, Expression, MypyFile, FuncDef, FuncItem, Decorator, RefExpr, Context, TypeInfo, ClassDef,
     Block, TypedDictExpr, NamedTupleExpr, AssignmentStmt, IndexExpr, TypeAliasExpr, NameExpr,
     CallExpr, NewTypeExpr, ForStmt, WithStmt, CastExpr, TypeVarExpr, TypeApplication, Lvalue,
-    TupleExpr, RevealTypeExpr, SymbolTableNode, Var, ARG_POS, OverloadedFuncDef, DepNode
+    TupleExpr, RevealTypeExpr, SymbolTableNode, Var, ARG_POS, OverloadedFuncDef
 )
 from mypy.types import (
     Type, Instance, AnyType, TypeOfAny, CallableType, TupleType, TypeVarType, TypedDictType,
@@ -32,6 +32,7 @@ from mypy.typevars import has_no_typevars
 from mypy.semanal_shared import PRIORITY_FORWARD_REF, PRIORITY_TYPEVAR_VALUES
 from mypy.subtypes import is_subtype
 from mypy.sametypes import is_same_type
+from mypy.scope import Scope
 import mypy.semanal
 
 
@@ -47,6 +48,7 @@ class SemanticAnalyzerPass3(TraverserVisitor):
         self.modules = modules
         self.errors = errors
         self.sem = sem
+        self.scope = Scope()
         # If True, process function definitions. If False, don't. This is used
         # for processing module top levels in fine-grained incremental mode.
         self.recurse_into_functions = True
@@ -61,31 +63,22 @@ class SemanticAnalyzerPass3(TraverserVisitor):
         self.is_typeshed_file = self.errors.is_typeshed_file(fnam)
         self.sem.cur_mod_id = file_node.fullname()
         self.cur_mod_node = file_node
-        # This variable tracks the current scope node: a module, a class definition,
-        # or a function. It is used to, e.g., correctly determine target of a fine-grained
-        # dependency.
-        self.cur_node = file_node  # type: DepNode
         self.sem.globals = file_node.names
         with experiments.strict_optional_set(options.strict_optional):
+            self.scope.enter_file(file_node.fullname())
             self.accept(file_node)
-
-    @contextmanager
-    def enter_scope(self, node: Union[MypyFile, ClassDef, FuncItem]) -> Iterator[None]:
-        old_node = self.cur_node
-        self.cur_node = node
-        try:
-            yield
-        finally:
-            self.cur_node = old_node
+            self.scope.leave()
 
     def refresh_partial(self, node: Union[MypyFile, FuncItem, OverloadedFuncDef]) -> None:
         """Refresh a stale target in fine-grained incremental mode."""
+        self.scope.enter_file(self.sem.cur_mod_id)
         if isinstance(node, MypyFile):
             self.recurse_into_functions = False
             self.refresh_top_level(node)
         else:
             self.recurse_into_functions = True
             self.accept(node)
+        self.scope.leave()
 
     def refresh_top_level(self, file_node: MypyFile) -> None:
         """Reanalyze a stale module top-level in fine-grained incremental mode."""
@@ -106,11 +99,12 @@ class SemanticAnalyzerPass3(TraverserVisitor):
     def visit_func_def(self, fdef: FuncDef) -> None:
         if not self.recurse_into_functions:
             return
+        self.scope.enter_function(fdef)
         self.errors.push_function(fdef.name())
-        with self.enter_scope(fdef):
-            self.analyze(fdef.type, fdef)
-            super().visit_func_def(fdef)
+        self.analyze(fdef.type, fdef)
+        super().visit_func_def(fdef)
         self.errors.pop_function()
+        self.scope.leave()
 
     def visit_overloaded_func_def(self, fdef: OverloadedFuncDef) -> None:
         if not self.recurse_into_functions:
@@ -121,6 +115,7 @@ class SemanticAnalyzerPass3(TraverserVisitor):
     def visit_class_def(self, tdef: ClassDef) -> None:
         # NamedTuple base classes are validated in check_namedtuple_classdef; we don't have to
         # check them again here.
+        self.scope.enter_class(tdef.info)
         if not tdef.info.is_named_tuple:
             types = list(tdef.info.bases)  # type: List[Type]
             for tvar in tdef.type_vars:
@@ -150,8 +145,8 @@ class SemanticAnalyzerPass3(TraverserVisitor):
             elif isinstance(tdef.analyzed, NamedTupleExpr):
                 self.analyze(tdef.analyzed.info.tuple_type, tdef.analyzed, warn=True)
                 self.analyze_info(tdef.analyzed.info)
-        with self.enter_scope(tdef):
-            super().visit_class_def(tdef)
+        super().visit_class_def(tdef)
+        self.scope.leave()
 
     def visit_decorator(self, dec: Decorator) -> None:
         """Try to infer the type of the decorated function.
@@ -371,7 +366,7 @@ class SemanticAnalyzerPass3(TraverserVisitor):
             self.check_for_omitted_generics(type)
             self.generate_type_patches(node, indicator, warn)
             if analyzer.aliases_used:
-                self.cur_mod_node.alias_deps[self.cur_node].update(analyzer.aliases_used)
+                self.cur_mod_node.alias_deps[self.scope.current_full_target()].update(analyzer.aliases_used)
 
     def analyze_types(self, types: List[Type], node: Node) -> None:
         # Similar to above but for nodes with multiple types.
@@ -381,7 +376,7 @@ class SemanticAnalyzerPass3(TraverserVisitor):
             type.accept(analyzer)
             self.check_for_omitted_generics(type)
             if analyzer.aliases_used:
-                self.cur_mod_node.alias_deps[self.cur_node].update(analyzer.aliases_used)
+                self.cur_mod_node.alias_deps[self.scope.current_full_target()].update(analyzer.aliases_used)
         self.generate_type_patches(node, indicator, warn=False)
 
     def generate_type_patches(self,
