+++ mypy/server/deps.py
@@ -92,7 +92,7 @@ from mypy.nodes import (
     ComparisonExpr, GeneratorExpr, DictionaryComprehension, StarExpr, PrintStmt, ForStmt, WithStmt,
     TupleExpr, ListExpr, OperatorAssignmentStmt, DelStmt, YieldFromExpr, Decorator, Block,
     TypeInfo, FuncBase, OverloadedFuncDef, RefExpr, SuperExpr, Var, NamedTupleExpr, TypedDictExpr,
-    LDEF, MDEF, GDEF, FuncItem, TypeAliasExpr, DepNode,
+    LDEF, MDEF, GDEF, FuncItem, TypeAliasExpr,
     op_methods, reverse_op_methods, ops_with_inplace_method, unary_op_methods
 )
 from mypy.traverser import TraverserVisitor
@@ -103,6 +103,7 @@ from mypy.types import (
 )
 from mypy.server.trigger import make_trigger
 from mypy.util import correct_relative_import
+from mypy.scope import Scope
 
 
 def get_dependencies(target: MypyFile,
@@ -146,7 +147,7 @@ class DependencyVisitor(TraverserVisitor):
     def __init__(self,
                  type_map: Dict[Expression, Type],
                  python_version: Tuple[int, int],
-                 alias_deps: 'DefaultDict[DepNode, Set[str]]') -> None:
+                 alias_deps: 'DefaultDict[str, Set[str]]') -> None:
         self.scope = Scope()
         self.type_map = type_map
         self.python2 = python_version[0] == 2
@@ -175,7 +176,7 @@ class DependencyVisitor(TraverserVisitor):
     def visit_mypy_file(self, o: MypyFile) -> None:
         self.scope.enter_file(o.fullname())
         self.is_package_init_file = o.is_package_init_file()
-        self.add_type_alias_deps(o)
+        self.add_type_alias_deps(self.scope.current_full_target())
         super().visit_mypy_file(o)
         self.scope.leave()
 
@@ -193,7 +194,7 @@ class DependencyVisitor(TraverserVisitor):
         if o.info:
             for base in non_trivial_bases(o.info):
                 self.add_dependency(make_trigger(base.fullname() + '.' + o.name()))
-        self.add_type_alias_deps(o)
+        self.add_type_alias_deps(self.scope.current_full_target())
         super().visit_func_def(o)
         self.scope.leave()
 
@@ -219,7 +220,7 @@ class DependencyVisitor(TraverserVisitor):
             self.add_type_dependencies(o.info.typeddict_type, target=make_trigger(target))
         # TODO: Add dependencies based on remaining TypeInfo attributes.
         super().visit_class_def(o)
-        self.add_type_alias_deps(o)
+        self.add_type_alias_deps(self.scope.current_full_target())
         self.is_class = old_is_class
         info = o.info
         for name, node in info.names.items():
@@ -534,15 +535,10 @@ class DependencyVisitor(TraverserVisitor):
 
     # Helpers
 
-    def add_type_alias_deps(self, o: DepNode) -> None:
-        """Add dependencies from type aliases to the current target."""
-        if o in self.alias_deps:
-            for alias in self.alias_deps[o]:
+    def add_type_alias_deps(self, target):
+        if target in self.alias_deps:
+            for alias in self.alias_deps[target]:
                 self.add_dependency(make_trigger(alias))
-                if isinstance(o, FuncItem):
-                    self.add_dependency(make_trigger(alias), make_trigger(o.fullname()))
-                elif isinstance(o, ClassDef):
-                    self.add_dependency(make_trigger(alias), make_trigger(o.fullname))
 
     def add_dependency(self, trigger: str, target: Optional[str] = None) -> None:
         """Add dependency from trigger to a target.
@@ -609,78 +605,6 @@ class DependencyVisitor(TraverserVisitor):
         typ = self.type_map.get(node)
         if typ:
             self.add_attribute_dependency(typ, '__iter__')
-
-
-class Scope:
-    """Track which target we are processing at any given time."""
-
-    def __init__(self) -> None:
-        self.module = None  # type: Optional[str]
-        self.classes = []  # type: List[TypeInfo]
-        self.function = None  # type: Optional[FuncDef]
-        # Number of nested scopes ignored (that don't get their own separate targets)
-        self.ignored = 0
-
-    def current_module_id(self) -> str:
-        assert self.module
-        return self.module
-
-    def current_target(self) -> str:
-        """Return the current target (non-class; for a class return enclosing module)."""
-        assert self.module
-        target = self.module
-        if self.function:
-            if self.classes:
-                target += '.' + '.'.join(c.name() for c in self.classes)
-            target += '.' + self.function.name()
-        return target
-
-    def current_full_target(self) -> str:
-        """Return the current target (may be a class)."""
-        assert self.module
-        target = self.module
-        if self.classes:
-            target += '.' + '.'.join(c.name() for c in self.classes)
-        if self.function:
-            target += '.' + self.function.name()
-        return target
-
-    def enter_file(self, prefix: str) -> None:
-        self.module = prefix
-        self.classes = []
-        self.function = None
-        self.ignored = 0
-
-    def enter_function(self, fdef: FuncDef) -> None:
-        if not self.function:
-            self.function = fdef
-        else:
-            # Nested functions are part of the topmost function target.
-            self.ignored += 1
-
-    def enter_class(self, info: TypeInfo) -> None:
-        """Enter a class target scope."""
-        if not self.function:
-            self.classes.append(info)
-        else:
-            # Classes within functions are part of the enclosing function target.
-            self.ignored += 1
-
-    def leave(self) -> None:
-        """Leave the innermost scope (can be any kind of scope)."""
-        if self.ignored:
-            # Leave a scope that's included in the enclosing target.
-            self.ignored -= 1
-        elif self.function:
-            # Function is always the innermost target.
-            self.function = None
-        elif self.classes:
-            # Leave the innermost class.
-            self.classes.pop()
-        else:
-            # Leave module.
-            assert self.module
-            self.module = None
 
 
 def get_type_triggers(typ: Type) -> List[str]:
